Divide cli.py into multiple files
