+++ pyproject.toml
@@ -45,7 +45,7 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-docli = 'doccano_api_client.cli:main'
+docli = 'doccano_api_client.cli.commands:main'
 
 [tool.black]
 line-length = 120
